autodoc_service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.0
python-docx==1.1.0
openpyxl==3.1.2
beautifulsoup4==4.12.3
lxml==5.2.1
python-multipart==0.0.7

# 개발/테스트 의존성 (선택사항)
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0